feat: add landing page and move user page to /user route
